# Slight speedup of step detection by smarter median computations