Use an online algorithm for precomputing the median deviations.

Gives a slight speedup for step detection with l1 norm. How much this is depends on the data.

Also optimize the Python overhead in memoization, which gives another small perf boost.

Also, add a benchmark suite (time to dogfood :)
```
   before     after       ratio
  [500a743a] [8e29109e]
-  543.27ms   480.81ms      0.89  step_detect.Simple.time_detect_regressions
-  238.54ms   167.30ms      0.70  step_detect.Simple.time_solve_potts_approx
```